Airframe has graduated from sbt-microcite
During the major upgrade sbt-microsite to 1.x, which introduced mdoc, we've tested several configurations that worked for our site. Then we decided to switch to use mdoc and Docusaurus to fully customize our web site as we wish: https://wvlet.org/airframe/

I'd like to thank you for developing this plugin. sbt-microsite has worked for us for years. As the number of documentation pages increases, our doc site is no longer at a micro-scale., so it would be a time to graduate.

Many thanks!